Refactor so app can be packaged properly
