+++ wowup-electron/package.json
@@ -92,7 +92,6 @@
     "node-gyp": "7.1.0",
     "npm-run-all": "4.1.5",
     "protractor": "~7.0.0",
-    "rxjs": "6.6.3",
     "spectron": "12.0.0",
     "ts-node": "9.0.0",
     "tslib": "2.0.1",
@@ -126,6 +125,7 @@
     "protobufjs": "6.10.1",
     "rimraf": "3.0.2",
     "rollbar": "2.19.3",
+    "rxjs": "6.6.3",
     "uuid": "8.3.1"
   }
 }
