+++ wowup-electron/src/app/components/titlebar/titlebar.component.ts
@@ -27,13 +27,10 @@ export class TitlebarComponent implements OnInit, OnDestroy {
   ) {
     const windowMaximizedSubscription = this.electronService.windowMaximized$.subscribe(
       (maximized) => {
+        console.log('subscription maximized:', maximized);
         this._ngZone.run(() => (this.isMaximized = maximized));
       }
     );
-
-    if (this.electronService.remote.getCurrentWindow().isMaximized) {
-      this.isMaximized = true;
-    }
 
     this._subscriptions = [windowMaximizedSubscription];
   }
