+++ wowup-electron/src/common/utils/number.utils.ts
